[#318] Refactor HTTP server request handling: extract `http_server_netconn_serve_handle_req` into a separate module for improved modularity and maintainability.
