+++ CMakeLists.txt
@@ -46,6 +46,8 @@ idf_component_register(
         src/http_server_mutex.h
         src/http_server_netconn_resp.c
         src/http_server_netconn_resp.h
+        src/http_server_netconn_serve_handle_req.c
+        src/http_server_netconn_serve_handle_req.h
         src/http_server_resp.c
         src/sta_ip_safe.c
         src/sta_ip_safe.h
